MEG upodate
Since we split the two steps of denoising, I decided to rename
aamod_meg_denoise_ICA to aamod_meg_denoise_ICA_1 (already in the
central), and the new module will be
aamod_meg_denoise_ICA_2_applytrajectory (not ready yet!). It also
reflects that the second one relies on the first one (similarly to
coreg_extended_1 and 2).