Support automated tests by refactoring examples.
1. aarecipe now creates a new seed aap_parameters_user.xml file if no
defaults file is specified. This put user-specific configuration outside of
version control (where it should be).

2. The example scripts all call aarecipe without a defaults file, so
that the above functionality makes it possible to run tests regardless
of site (as long as the data is publicly available).

3. aa_test/aatest - new test function for running tests on the above
examples in an automated manner.

4. Re-organised the examples directory for clarity and to make it
possible to run particular test flavours (e.g., all CBU-specific tests
are under the cbu folder). Added light-weight documentation for this.

5. Big tidy up of legacy examples.

6. Relocated the task lists to the examples directory, and renamed to
always match the user script it goes with. Renamed
aa_recipes_and_parametersets aa_parametersets to reflect this change
(this directory only contains defaults now).