removes 'obj' output args from methods of qsub
All queue processor classes including qsub are handle classes, meaning that variables instantiated with that class are pointers to an object that is shared among all workspaces. Copies of that variable are all pointers to the same object. So, if any class method changes properties of its own instance, these changes will be 'global' in the sense that they are instantly effective also outside the method's workspace, obviating the need to return the class instance from any of the class' methods.